Fix comparison warning because of OpenMP

+++ libeasea/include/algorithms/moea/Csigma.h
@@ -114,7 +114,7 @@ typename Csigma<TIndividual, TRandom>::TPopulation Csigma<TIndividual, TRandom>:
 #ifdef USE_OPENMP
     EASEA_PRAGMA_OMP_PARALLEL
 #endif
-        for (int i = 0; i < offspring.size(); ++i)
+        for (int i = 0; i < static_cast<int>(offspring.size()); ++i)
         {
                 TI &child = offspring[i];
